Split shell computer accessibility tests
